+++ core/src/main/resources/db/migration/V3__create-table-price.sql
@@ -1,7 +1,7 @@
 CREATE TABLE price (
     Id INT PRIMARY KEY AUTO_INCREMENT,
     assetsId INT NOT NULL,
-    unitPrice DECIMAL(10,2) NOT NULL,
+    unitPrice INT NOT NULL,
     priceDate DATETIME NOT NULL,
 
     FOREIGN KEY (assetsId) REFERENCES assets(id)
